+++ tests/test_parasail.rs
@@ -2,39 +2,46 @@ use parasail_rs::{Aligner, Matrix, Profile};
 use std::thread;
 
 #[test]
-pub fn matrix_construction() {
+pub fn matrix_construction() -> Result<(), Box<dyn std::error::Error>> {
     // default matrix
     Matrix::default();
 
     // custom matrix
-    let mut matrix = Matrix::create(b"ACGT", 3, -2);
+    let mut matrix = Matrix::create(b"ACGT", 3, -2)?;
+    println!("Matrix:\n {}", matrix);
+    matrix.set_value(2, 2, 100);
+    println!("Matrix:\n {}", matrix);
 
     // from name
-    Matrix::from("blosum62");
+    let blosum62 = Matrix::from("blosum62")?;
+
+    // convert to PSSM
+    let _blosum62_pssm = blosum62.to_pssm(b"ACGT");
 
     // square matrix from file
-    Matrix::from_file("./tests/square.txt");
+    Matrix::from_file("./tests/square.txt")?;
 
     // pssm from file
-    Matrix::from_file("./tests/pssm.txt");
+    Matrix::from_file("./tests/pssm.txt")?;
 
     // PSSM
-    // let values = vec![1, 2, 3, 4];
-    // let rows = vec![b'A', b'C'];
-    // Matrix::create_pssm("AC", values, rows);
+    let pssm_alphabet = "abcdef";
+    let values = vec![1, 2, 3, 4, 5, 6, 7, 8];
+    let rows = 2;
+    Matrix::create_pssm(pssm_alphabet, values, rows)?;
 
-    // convert square to pssm
-    //
-    matrix.set_value(0, 0, 5);
+    Ok(())
 }
 
 #[test]
-pub fn profile_construction() {
+pub fn profile_construction() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ATGGCACTATAA";
-    Profile::new(query, false, &Matrix::default());
+    Profile::new(query, false, &Matrix::default())?;
 
     // with stats
-    Profile::new(query, true, &Matrix::default());
+    Profile::new(query, true, &Matrix::default())?;
+
+    Ok(())
 }
 
 #[test]
@@ -55,11 +62,11 @@ pub fn aligner_construction() {
 }
 
 #[test]
-pub fn global_alignment() {
+pub fn global_alignment() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
 
@@ -70,14 +77,16 @@ pub fn global_alignment() {
     assert!(!result.is_local());
     assert!(!result.is_semi_global());
     assert!(result.is_striped());
+
+    Ok(())
 }
 
 #[test]
-pub fn semi_global_alignment() {
+pub fn semi_global_alignment() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().semi_global().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
 
@@ -88,14 +97,16 @@ pub fn semi_global_alignment() {
     assert!(!result.is_local());
     assert!(result.is_semi_global());
     assert!(result.is_striped());
+
+    Ok(())
 }
 
 #[test]
-pub fn local_alignment() {
+pub fn local_alignment() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().local().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
 
@@ -106,6 +117,8 @@ pub fn local_alignment() {
     assert!(result.is_local());
     assert!(!result.is_semi_global());
     assert!(result.is_striped());
+
+    Ok(())
 }
 
 #[test]
@@ -113,7 +126,7 @@ pub fn global_with_stats() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().use_stats().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
     let n_matches = result.get_matches()?;
@@ -130,7 +143,7 @@ pub fn semi_global_with_stats() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().semi_global().use_stats().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
     let n_matches = result.get_matches()?;
@@ -147,7 +160,7 @@ pub fn local_with_stats() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().local().use_stats().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
 
     let checks = query.len() as i32;
     let n_matches = result.get_matches()?;
@@ -167,7 +180,7 @@ pub fn score_table() -> Result<(), Box<dyn std::error::Error>> {
     let aligner = Aligner::new().use_table().build();
     let default_score = 1;
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_table());
     assert!(!result.is_stats());
     assert!(!result.is_stats_table());
@@ -176,7 +189,7 @@ pub fn score_table() -> Result<(), Box<dyn std::error::Error>> {
     // one-off alignment with stats
     let aligner = Aligner::new().use_stats().use_table().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats());
     assert!(result.is_stats_table());
     assert!(result.is_table());
@@ -184,12 +197,12 @@ pub fn score_table() -> Result<(), Box<dyn std::error::Error>> {
 
     // alignment with profile, without stats
     let custom_score = 3;
-    let matrix = Matrix::create(b"ACGT", custom_score, -2);
-    let profile = Profile::new(query, false, &matrix);
+    let matrix = Matrix::create(b"ACGT", custom_score, -2)?;
+    let profile = Profile::new(query, false, &matrix)?;
 
     let aligner_w_profile = Aligner::new().profile(profile).use_table().build();
 
-    let result_w_profile = aligner_w_profile.align(None, reference);
+    let result_w_profile = aligner_w_profile.align(None, reference)?;
 
     assert!(result_w_profile.is_table());
     assert!(!result_w_profile.is_stats());
@@ -197,14 +210,14 @@ pub fn score_table() -> Result<(), Box<dyn std::error::Error>> {
     assert_eq!(result_w_profile.get_score_table()?, custom_score);
 
     // // alignment with profile, with stats
-    let profile = Profile::new(query, true, &matrix);
+    let profile = Profile::new(query, true, &matrix)?;
     let aligner_w_profile = Aligner::new()
         .profile(profile)
         .use_stats()
         .use_table()
         .build();
 
-    let result_w_profile = aligner_w_profile.align(None, reference);
+    let result_w_profile = aligner_w_profile.align(None, reference)?;
 
     assert!(result_w_profile.is_stats());
     assert!(result_w_profile.is_stats_table());
@@ -222,7 +235,7 @@ pub fn matches_table() -> Result<(), Box<dyn std::error::Error>> {
     let aligner = Aligner::new().use_table().use_stats().build();
     let default_score = 1;
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_table());
     assert!(result.is_stats());
     assert!(result.is_stats_table());
@@ -238,7 +251,7 @@ pub fn similar_table() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_table().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_table());
     assert!(result.is_stats());
     assert!(result.is_stats_table());
@@ -254,7 +267,7 @@ pub fn length_table() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_table().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_table());
     assert!(result.is_stats());
     assert!(result.is_stats_table());
@@ -270,7 +283,7 @@ pub fn score_row() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -286,7 +299,7 @@ pub fn matches_row() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -302,7 +315,7 @@ pub fn similar_row() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -318,7 +331,7 @@ pub fn length_row() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -334,7 +347,7 @@ pub fn score_col() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -350,7 +363,7 @@ pub fn match_col() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -366,7 +379,7 @@ pub fn similar_col() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -382,7 +395,7 @@ pub fn length_col() -> Result<(), Box<dyn std::error::Error>> {
     let reference = b"ACGT";
     let aligner = Aligner::new().use_last_rowcol().use_stats().build();
 
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_stats_rowcol());
     assert!(result.is_stats());
     assert!(!result.is_stats_table());
@@ -396,7 +409,7 @@ pub fn trace_table() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().use_trace().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     assert!(result.is_trace());
     println!("Trace table: {:?}", result.get_trace_table()?);
 
@@ -432,7 +445,7 @@ pub fn get_traceback_strings() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().use_trace().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     let traceback = result.get_traceback_strings(query, reference)?;
     println!("Query:     {}", traceback.query);
     println!("           {}", traceback.comparison);
@@ -446,8 +459,8 @@ pub fn print_traceback() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().use_trace().build();
-    let result = aligner.align(Some(query), reference);
-    result.print_traceback(query, reference)?;
+    let result = aligner.align(Some(query), reference)?;
+    result.print_traceback(query, reference);
 
     Ok(())
 }
@@ -457,7 +470,7 @@ pub fn get_cigar() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let aligner = Aligner::new().use_trace().build();
-    let result = aligner.align(Some(query), reference);
+    let result = aligner.align(Some(query), reference)?;
     let cigar_string = result.get_cigar(query, reference)?;
 
     println!("CIGAR: {}", cigar_string);
@@ -466,11 +479,11 @@ pub fn get_cigar() -> Result<(), Box<dyn std::error::Error>> {
 }
 
 #[test]
-pub fn global_with_profile() {
+pub fn global_with_profile() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let matrix = Matrix::default();
-    let profile = Profile::new(query, true, &matrix);
+    let profile = Profile::new(query, true, &matrix)?;
 
     let aligner = Aligner::new()
         .profile(profile)
@@ -478,20 +491,22 @@ pub fn global_with_profile() {
         .striped()
         .build();
 
-    let result = aligner.align(None, reference);
+    let result = aligner.align(None, reference)?;
     assert!(result.is_global());
     assert!(result.is_striped());
     assert!(result.is_stats());
     assert!(!result.is_local());
     assert!(!result.is_semi_global());
+
+    Ok(())
 }
 
 #[test]
-pub fn semi_global_with_profile() {
+pub fn semi_global_with_profile() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let matrix = Matrix::default();
-    let profile = Profile::new(query, true, &matrix);
+    let profile = Profile::new(query, true, &matrix)?;
 
     let aligner = Aligner::new()
         .profile(profile)
@@ -500,20 +515,22 @@ pub fn semi_global_with_profile() {
         .semi_global()
         .build();
 
-    let result = aligner.align(None, reference);
+    let result = aligner.align(None, reference)?;
     assert!(result.is_semi_global());
     assert!(result.is_striped());
     assert!(result.is_stats());
     assert!(!result.is_local());
     assert!(!result.is_global());
+
+    Ok(())
 }
 
 #[test]
-pub fn local_with_profile() {
+pub fn local_with_profile() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let reference = b"ACGT";
     let matrix = Matrix::default();
-    let profile = Profile::new(query, true, &matrix);
+    let profile = Profile::new(query, true, &matrix)?;
 
     let aligner = Aligner::new()
         .profile(profile)
@@ -522,30 +539,40 @@ pub fn local_with_profile() {
         .local()
         .build();
 
-    let result = aligner.align(None, reference);
+    let result = aligner.align(None, reference)?;
     assert!(result.is_local());
     assert!(result.is_striped());
     assert!(result.is_stats());
     assert!(!result.is_global());
     assert!(!result.is_semi_global());
+
+    Ok(())
 }
 
 #[test]
-pub fn multithread_global_alignment() {
+pub fn multithread_global_alignment() -> Result<(), Box<dyn std::error::Error>> {
     let query = b"ACGT";
     let refs = vec![b"ACGT", b"ACGT"];
     let matrix = Matrix::default();
-    let profile = Profile::new(query, true, &matrix);
+    let profile = Profile::new(query, true, &matrix)?;
 
     let aligner = Aligner::new().profile(profile).use_stats().build();
 
     thread::spawn(move || {
         for reference in refs {
-            let result = &aligner.align(None, reference);
-            let score = result.get_score();
-            assert_eq!(score, query.len() as i32);
+            match &aligner.align(None, reference) {
+                Ok(result) => {
+                    let score = result.get_score();
+                    assert_eq!(score, query.len() as i32);
+                }
+                Err(e) => {
+                    println!("Alignment Error: {}", e);
+                }
+            }
         }
     })
     .join()
     .unwrap();
+
+    Ok(())
 }